requirements/production.txt
dj-database-url==0.5.0
Django==3.0.5
django-filter==2.2.0
djangorestframework==3.11.0
drf-yasg==1.17.1
prettyconf==2.1.0
psycopg2==2.8.5
whitenoise==5.0.1